refactor(api): consolidate queries from workflows
