+++ apps/api/src/lib/workflows/posts.ts
@@ -1,6 +1,6 @@
 import { generatePost } from "@api/lib/gemini/generate-post";
-import { getCarsByMonth } from "@api/lib/workflows/queries/cars";
-import { getCoeForMonth } from "@api/lib/workflows/queries/coe";
+import { getCarsAggregatedByMonth } from "@api/queries";
+import { getCoeForMonth } from "@api/queries/coe";
 import type { WorkflowContext } from "@upstash/workflow";
 
 export const generateCarPost = async (
@@ -8,7 +8,7 @@ export const generateCarPost = async (
   month: string,
 ) => {
   // Get car data from database
-  const cars = await getCarsByMonth(month);
+  const cars = await getCarsAggregatedByMonth(month);
   const data = cars.map((row) => JSON.stringify(row));
 
   return generatePost(context, {
