+++ apps/api/src/queries/index.ts
@@ -1,1 +1,2 @@
 export * from "./cars";
+export * from "./coe";
